Function ic_registry_common::bootstrap_registry_data_provider::is_valid_initial_update

pub fn is_valid_initial_update(
    update: &(Vec<RegistryTransportRecord>, RegistryVersion)
) -> bool

Tests whether the update received from the initial data provider is a legal first version of the registry, i.e. only contains records with version 1 and the latest version is 1.