Module ic_registry_common::local_store

Structs

KeyMutation

LocalStoreImpl

Traits

LocalStore

LocalStoreReader

LocalStoreWriter

Type Definitions

Changelog
A changelog is a sequence of ChangelogEntries. Applied consecutively to a registry, the registry will be at version v + l, if v is the previous latest version and l is the length of the Changelog.

ChangelogEntry
A ChangelogEntry is a list of mutations that, when applied to a registry at version v, produce the registry at version v+1.