Trait ic_registry_common::local_store::LocalStoreWriter

pub trait LocalStoreWriter: Send + Sync {
    fn store(&self, version: RegistryVersion, v: ChangelogEntry) -> Result<()>;

    fn clear(&self) -> Result<()>;

    fn update_certified_time(&self, unix_epoch_nanos: u64) -> Result<()>;
}

Required methods

fn store(&self, version: RegistryVersion, v: ChangelogEntry) -> Result<()>

Store the changelog at the given version.

Preconditions: (1) For a given version > 1, a changelog_entry for version-1 must exist in the store. (2) The given change log entry must be nonempty list of KeyMutations.

fn clear(&self) -> Result<()>

Clears the Local Store.

Note: This clears registry versions, stored in directories, but not the certified timestamp file in the root of the local store.

fn update_certified_time(&self, unix_epoch_nanos: u64) -> Result<()>

Update the locally stored certified time to unix_epoch_nanos.

Loading content...