Type Definition ic_registry_common::local_store::ChangelogEntry

type ChangelogEntry = Vec<KeyMutation>;

A ChangelogEntry is a list of mutations that, when applied to a registry at version v, produce the registry at version v+1.